refactor: move field-tooltips to lib/, add @base-ui comment
Move field-tooltips.ts from data/ to lib/ for consistency with
existing static lookup modules (cefr-colors, languages). Add a
comment explaining why FieldTooltip uses @base-ui directly.
