+++ frontend/src/components/FieldTooltip.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { describe, it, expect } from 'vitest'
-import { FieldTooltip } from './FieldTooltip'
+import { FieldTooltip } from '@/components/FieldTooltip'
 
 describe('FieldTooltip', () => {
   it('renders info icon for a known field', () => {
